requirements.txt
tensorflow-gpu
opencv-python
Pillow
keras
wsq
